Move code to internal
