+++ cmd/cli/main.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/cbebe/work-tracker/pkg/work"
+	"github.com/cbebe/work-tracker/internal/work"
 )
 
 func printUsage() {
