+++ internal/web/work_controller.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"text/template"
 
-	"github.com/cbebe/work-tracker/pkg/work"
+	"github.com/cbebe/work-tracker/internal/work"
 )
 
 type WorkController struct {
